feat: fill the cell for box-drawing + block glyphs
the font's box-drawing glyphs are only ~1 em tall, so with line_height > 1 they left vertical seams and box borders looked broken. draw the box-drawing lines/corners/tees and the block + shade elements programmatically to the full cell instead, so borders connect at any line height; rounded corners keep a quarter-arc. unsupported glyphs (double, dashed, quadrants) fall back to the font. also drop a redundant isize cast in explorer_dir_for so clippy stays clean.

+++ src/win.rs
@@ -437,7 +437,7 @@ pub fn explorer_dir_for(hwnd: isize) -> Option<String> {
             let Ok(wb) = disp.cast::<IWebBrowser2>() else {
                 continue;
             };
-            let matches = wb.HWND().map(|w| w.0 as isize == hwnd).unwrap_or(false);
+            let matches = wb.HWND().map(|w| w.0 == hwnd).unwrap_or(false);
             if !matches {
                 continue;
             }
